Refactor into less files
