+++ src/Exercism.Representers.CSharp/Exercism.Representers.CSharp.csproj
@@ -3,12 +3,10 @@
     <PropertyGroup>
         <TargetFramework>net7.0</TargetFramework>
         <OutputType>Exe</OutputType>
-        <PublishAot>true</PublishAot>
     </PropertyGroup>
     
     <ItemGroup>
         <PackageReference Include="Humanizer.Core" Version="2.14.1" />
-        <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.4.0" />
         <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="4.4.0" />
     </ItemGroup>
