--- src/Exercism.Representers.CSharp/Options.cs
@@ -1,3 +1,0 @@
-namespace Exercism.Representers.CSharp;
-
-internal record Options(string Slug, string InputDirectory, string OutputDirectory);
